refactor:removed code from home component into several separate files
