+++ src/components/ProtectedRoute.tsx
@@ -1,14 +1,16 @@
 import { Navigate } from 'react-router-dom';
 import { ReactNode } from 'react';
 import { useAuth } from '../hooks/useAuth';
+import { useTranslation } from 'react-i18next';
 
 type Props = { children: ReactNode };
 
 export default function ProtectedRoute({ children }: Props) {
   const { user, loading } = useAuth();
+  const { t } = useTranslation();
 
   if (loading) {
-    return <div className="p-6 text-center">Завантаження…</div>;
+    return <div className="p-6 text-center">{t('Loading')}</div>;
   }
   if (!user) {
     return <Navigate to="/" replace />;
