+++ src/components/pages/index.tsx
@@ -1,0 +1,5 @@
+import HistoryPage from './HistoryPage';
+import MedicationsPage from './MedicationsPage';
+import PatientsPage from './PatientsPage';
+
+export { HistoryPage, MedicationsPage, PatientsPage };
