+++ src/landing/components/index.tsx
@@ -1,0 +1,7 @@
+import About from './About';
+import Features from './Features';
+import Footer from './Footer';
+import Hero from './Hero';
+import Navigation from './Navigation';
+import JoinUs from './JoinUs';
+export { About, Features, Footer, Hero, Navigation, JoinUs };
