+++ src/locales/en.json
@@ -395,5 +395,6 @@
         "content": "If you have questions about our privacy policy or data processing, contact us at: privacy@axelspills.com"
       }
     }
-  }
+  },
+  "Loading": "Loading..."
 }
