+++ src/locales/uk.json
@@ -395,5 +395,6 @@
         "content": "Якщо у вас є питання щодо нашої політики конфіденційності або обробки ваших даних, зв'яжіться з нами за адресою: privacy@axelspills.com"
       }
     }
-  }
+  },
+  "Loading": "Завантаження..."
 }
